+++ src/main/kotlin/fr/heta__h/squ_abyssal_bloom/config/ServerConfigData.kt
@@ -6,28 +6,28 @@ import net.minecraft.network.codec.StreamCodec
 
 data class ServerConfigData(
     val strictBarnacleSpawning: Boolean,
-    val shallowDeepBoundary: Float,
-    val deepAbyssalBoundary: Float,
+    val shallowDeepBoundary: Double,
+    val deepAbyssalBoundary: Double,
 ) {
     companion object {
         val STREAM_CODEC: StreamCodec<RegistryFriendlyByteBuf, ServerConfigData> = StreamCodec.composite(
             ByteBufCodecs.BOOL, ServerConfigData::strictBarnacleSpawning,
-            ByteBufCodecs.FLOAT, ServerConfigData::shallowDeepBoundary,
-            ByteBufCodecs.FLOAT, ServerConfigData::deepAbyssalBoundary,
+            ByteBufCodecs.DOUBLE, ServerConfigData::shallowDeepBoundary,
+            ByteBufCodecs.DOUBLE, ServerConfigData::deepAbyssalBoundary,
             ::ServerConfigData
         )
 
         fun fromSpec(): ServerConfigData = ServerConfigData(
             strictBarnacleSpawning = ModServerConfig.STRICT_BARNACLE_SPAWNING.get(),
-            shallowDeepBoundary = ModServerConfig.SHALLOW_DEEP_BOUNDARY.get().toFloat(),
-            deepAbyssalBoundary = ModServerConfig.DEEP_ABYSSAL_BOUNDARY.get().toFloat(),
+            shallowDeepBoundary = ModServerConfig.SHALLOW_DEEP_BOUNDARY.get(),
+            deepAbyssalBoundary = ModServerConfig.DEEP_ABYSSAL_BOUNDARY.get(),
         )
     }
 
     fun applyToSpec() {
         ModServerConfig.STRICT_BARNACLE_SPAWNING.set(strictBarnacleSpawning)
-        ModServerConfig.SHALLOW_DEEP_BOUNDARY.set(shallowDeepBoundary.toDouble())
-        ModServerConfig.DEEP_ABYSSAL_BOUNDARY.set(deepAbyssalBoundary.toDouble())
+        ModServerConfig.SHALLOW_DEEP_BOUNDARY.set(shallowDeepBoundary)
+        ModServerConfig.DEEP_ABYSSAL_BOUNDARY.set(deepAbyssalBoundary)
         ModServerConfig.SPEC.save()
     }
 }
